refactor(server_stdio): relocate graph/concept tools to services/stdio_graph_tools.py
Option 2 of docs/current/server-stdio-module-split.md, deferred until
Option 1 shipped and ran stable (merged via PR #88). Moves the 5 graph
tools (marm_graph_index, marm_code_lookup, marm_graph_trace,
marm_graph_architecture, marm_graph_impact) and the 2 concept tools
(marm_concept_build, marm_concept_recall), plus their private
_graph_available/_graph_unavailable helpers, out of server_stdio.py
into a new services/stdio_graph_tools.py module. They register onto a
FastMCP instance now shared via a new core/stdio_mcp_app.py module,
avoiding a circular import between server_stdio.py and the new tool
module.

server_stdio.py drops from 632 to 339 lines, under the project's
400-line check-file-length.py threshold. The 7 moved functions are
re-imported into server_stdio.py so existing `stdio.marm_graph_index`
style access still resolves.

Extraction was script-based index-slicing from git history, not
retyped, and integrity-checked byte-for-byte against an independent
re-slice of the original file (both the moved body and the untouched
tail).

Updated tests/test_stdio_transport.py: 5 monkeypatch.setattr(stdio, ...)
calls patched names (_run_recall, _marm_concept_build_endpoint,
graph_router) that now live in stdio_graph_tools's own module globals,
not server_stdio's -- redirected each to target the new module.
Confirmed the old attribute paths raise AttributeError on server_stdio
now, proving the redirects are load-bearing rather than cosmetic.

Full suite: 626 passed, 29 skipped. Live smoke test confirmed all 14
tools still register on the shared mcp instance and both a moved tool
and an unmoved tool execute correctly end to end.
